+++ CHANGELOG.rst
@@ -1,5 +1,8 @@
 100.2.9-dev (Unreleased)
 ------------------------
+* feat: improved ldap integration with pas.plugins.ldap
+* feat: added migrate_ldap script
+* feat: added ldap_group_roles plugin to manage groups and users role assignments
 * fix: include Products.PluggableAuthService in configure.zcml to fix AttributeError: @@csrf_token in some ZMI pages
 * feat: added reportek_modification_date to avoid relying on Zope's _p_mtime for bobobase_modification_time
 * fix: sonarqube fixes